requirements.txt

pandas==1.4.2
streamlit ==1.10.0
streamlit-folium==0.6.13
streamlit-lottie == 0.0.3 
streamlit-aggrid == 0.2.3.post2 
geopy == 2.2.0
streamlit-option-menu == 0.3.2
PulP == 2.6.0



